Changed the definition of height
This commit changes the way the height is read by lammps. Before, the height simply defined the stiffness of the negative force. This was an error, because the height was given in energy units. Now the height really gives the height of the potential at the center of the trap. In the future, it would be a good idea to mimic the old behavior when the height is given in stiffness units.